Feat: arroyostack files separated, ready for assembling
