+++ src/api/arroyostack_routes_admin_models/models/post_model.py
@@ -1,0 +1,29 @@
+from api.db import db
+
+class PostModel(db.Model):
+    __tablename__ = "post"
+    id = db.Column(db.BigInteger, default=lambda: uuid.uuid4().int >> (
+        128 - 32), primary_key=True)
+    title = db.Column(db.String(300), unique=False, nullable=False)
+    body = db.Column(db.String(3000), unique=False, nullable=False)
+    image = db.Column(db.String(3000), unique=False, nullable=False)
+    category_enum = db.Enum('Salud', 'Bienestar',
+                            'Belleza', name='category_enum')
+    author = db.Column(db.String(200), unique=False, nullable=False)
+    category = db.Column(category_enum, nullable=False)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+
+    def __repr__(self):
+        return f'<Post: {self.id} {self.email}>'
+
+    def serialize(self):
+        return {
+            "id": self.id,
+            "title": self.title,
+            "body": self.body,
+            "date": self.date,
+            "image": self.image,
+            "category": self.category,
+            "author": self.author,
+            "category": self.category
+        }
